(improvement)(Chat) Split chatService into chatManageService and chatService
